+++ R/spsUIcollections.R
@@ -24,9 +24,9 @@
 #' @param texts vector of labels under each image
 #' @param hrefs vector of links when each image is clicked
 #' @param image_frame_size integer, 1-12, this controls width. How large is each
-#' image. 12 is the whole width of the screen and 1 is 1/12 of the screen. Consider
-#' numbers than can fully divide 12, like 1, 2, 3, 4, 6 or 12 (if you want only 1 image
-#' per row).
+#' image. 12 is the whole width of the parent container and 1 is 1/12 of the container. Consider
+#' numbers that can be fully divided by 12, like 1 (12 per row), 2 (6 per row),
+#' 3 (4 per row), 4 (3 per row), 6 (1 per row)or 12 (if you want only 1 image per row).
 #' @param images a vector of image sources, can be online URLs or local resource paths.
 #' @param enlarge bool,  when click on the image, enlarge
 #' it? If enlarge is enabled, click the photo will enlarge instead of jump to the link.
@@ -156,18 +156,20 @@ gallery <- function(texts,
             gallery_div
         ),
         if(enlarge && enlarge_method == "modal") {
-          singleton(
-            div(
-              id = "sps-gallery-modal",class = "gallery-modal",
-              onclick = "galModalClose()",
-              tags$span(
-                class="gallery-modal-close",
-                "X"
-              ),
-              tags$img(id="sps-gallery-modal-content", class="gallery-modal-content"),
-              div(class = "gallery-caption")
-            )
-          )
+          tags$script(HTML(
+          "
+          (function(){
+            if ($('#sps-gallery-modal').length) return ;
+            const gal_modal  =
+            `<div id='sps-gallery-modal' class='gallery-modal' onclick='galModalClose()' style='display: none;'>
+              <span class='gallery-modal-close'>X</span>
+              <img id='sps-gallery-modal-content' class='gallery-modal-content'/>
+              <div class='gallery-caption'></div>
+            </div>`;
+            $(document.body).append(gal_modal);
+          })()
+          "
+          ))
         } else "",
         tags$script(glue('fixGalHeight("{Id}")')),
         spsDepend("basic")